Moved useEffect and firestore into polls components
