+++ src/screens/userScreens/ManagementScreen.tsx
@@ -1,5 +1,5 @@
 // React ⚛
-import React, {useState, useEffect, useCallback} from 'react';
+import React, {useState, useCallback} from 'react';
 
 // React Native 📱
 import {View, Platform, SafeAreaView, ScrollView} from 'react-native';
@@ -13,55 +13,14 @@ import EventsManagement from '../../components/screenComponents/EventsManagement
 // StyleSheets 🌈
 import {CoreStyleSheet} from '../../StyleSheets/CoreDesignStyleSheet';
 
-// FireStore 🔥
-import {getLofftPolls} from '../../api/firebase/fireStoreActions';
-import firestore from '@react-native-firebase/firestore';
-
 const ManagementScreen = ({navigation}: any) => {
   // User Hooks
   const [image, setImage]: any = useState('');
   // Hooks
   const [calendarActive, setCalendarActive] = useState(true);
-  const [polls, setPolls] = useState([]);
-  const [pastPolls, setPastPolls] = useState([]);
-  const [todayDate] = useState(new Date());
 
   const buttonToggle = useCallback(toggled => {
     setCalendarActive(toggled);
-  }, []);
-
-  useEffect(() => {
-    const pollsData = async () => {
-      setPolls([]);
-      const allPolls = await getLofftPolls();
-      const currentPolls = allPolls.filter(
-        poll =>
-          (poll.data().deadline &&
-            new Date(poll.data().deadline.seconds * 1000) > todayDate) ||
-          !poll.data().deadline,
-      );
-      const oldPolls = allPolls.filter(
-        poll =>
-          poll.data().deadline &&
-          new Date(poll.data().deadline.seconds * 1000) < todayDate,
-      );
-      setPolls(currentPolls);
-      setPastPolls(oldPolls);
-    };
-
-    const subscriber = firestore()
-      .collection('Managements')
-      .doc('B7vxlFYgNpnYPOT7eMfO')
-      .collection('Polls')
-      .onSnapshot(snapShot => {
-        snapShot.docChanges().forEach(async change => {
-          if (change.type === 'added' || change.type === 'removed') {
-            pollsData();
-            console.log(change.type);
-          }
-        });
-      });
-    return () => subscriber();
   }, []);
 
   return (
@@ -75,7 +34,7 @@ const ManagementScreen = ({navigation}: any) => {
           <HeaderBar title="Management" image={image} />
 
           <ToggleBar
-            optionA="Calendar"
+            optionA="Events & Calendar"
             optionB="Polls"
             dashboard={buttonToggle}
           />
@@ -83,11 +42,7 @@ const ManagementScreen = ({navigation}: any) => {
           {calendarActive ? (
             <EventsManagement navigation={navigation} />
           ) : (
-            <PollsManagement
-              navigation={navigation}
-              pastPolls={pastPolls}
-              polls={polls}
-            />
+            <PollsManagement navigation={navigation} />
           )}
         </ScrollView>
       </SafeAreaView>
